+++ src/components/flyout/flyout.js
@@ -46,6 +46,7 @@ export class EuiFlyout extends Component {
 
     const flyoutContent = (
       <div
+        role="dialog"
         ref={node => { this.flyout = node; }}
         className={classes}
         tabIndex={0}
@@ -68,6 +69,8 @@ export class EuiFlyout extends Component {
     return (
       <span>
         {optionalOverlay}
+        {/* Trap focus even when ownFocus={false}, otherwise closing the flyout won't return focus
+            to the originating button */}
         <FocusTrap
           focusTrapOptions={{
             fallbackFocus: () => this.flyout,